+++ src/Farmer/Builders/Builders.ContainerRegistry.fs
@@ -43,15 +43,18 @@ type ContainerRegistryBuilder() =
 
     [<CustomOperation "name">]
     /// Sets the name of the Azure Container Registry instance.
-    member _.Name (state:ContainerRegistryConfig, name) = { state with Name = ResourceName name }
+    member _.Name (state: ContainerRegistryConfig, name: ResourceName) =
+        { state with Name = ContainerRegistryValidation.ContainerRegistryName.Create(name).OkValue.ResourceName }
+    member this.Name(state: ContainerRegistryConfig, name: string) = this.Name(state, ResourceName name)
+
 
     [<CustomOperation "sku">]
     /// Sets the name of the SKU/Tier for the Container Registry instance.
-    member _.Sku (state:ContainerRegistryConfig, sku) = { state with Sku = sku }
+    member _.Sku (state: ContainerRegistryConfig, sku) = { state with Sku = sku }
 
     [<CustomOperation "enable_admin_user">]
     /// Enables the admin user on the Azure Container Registry.
-    member _.EnableAdminUser (state:ContainerRegistryConfig) = { state with AdminUserEnabled = true }
+    member _.EnableAdminUser (state: ContainerRegistryConfig) = { state with AdminUserEnabled = true }
     interface ITaggable<ContainerRegistryConfig> with member _.Add state tags = { state with Tags = state.Tags |> Map.merge tags }
 
 let containerRegistry = ContainerRegistryBuilder()