+++ RELEASE_NOTES.md
@@ -1,8 +1,32 @@
 Release Notes
 =============
 
-## 1.6.31
+## 1.6.35
 * Azure CLI: Escape parameters passed to the az deployment command (breaking change). Any previously escaped parameters need to be unescaped before passing to the tryValidate, tryWhatIf, tryExecute, whatIf and execute functions.
+
+## 1.6.34
+* CLI: Include `--overwrite true` option when executing `az storage blob upload-batch` with Azure CLI 2.34.0 and above.
+* Container Groups: Deploy container groups to a specific zone.
+* Container Groups: Diagnostics support to send logs to a Log Analytics workspace.
+* Container Groups: Support for attaching to subnets directly without requiring a network profile.
+
+## 1.6.33
+* Container Groups: Specify DNS nameservers and search domains.
+* Container Registry: Adds name validation
+* DNS: Add support for private DNS zones and records
+* PostgreSQL: Added possibility to set vnet rules for PostgreSQL.
+* WebApps: Support virtual applications with `add_virtual_application`/`add_virtual_application_preloaded`
+
+## 1.6.32
+* DiagnosticSettings now supports resources that contain multiple segments e.g. SQL Databases.
+* ContainerApps now use the updated resource name (Microsoft.App instead of Microsoft.Web).
+* Updated documentation on main page from `Writer.quickDeploy` to `Writer.quickWrite`
+
+## 1.6.31
+* WebApps: Fix flakey deployments of web apps with multiple custom domains.
+* Deployments: Fix `ResourceId` generation when using a resource with a template.
+* AzureFirewall: Supports availability zones
+* WebApps/Functions: Add support for vnet integration
 
 ## 1.6.30
 * WebApps/Functions: Specify connection string types